requirements.txt
pandas~=2.3.2
numpy~=2.3.3
scikit-learn~=1.7.2
python-dotenv~=1.1.1
loguru~=0.7.3
pymongo~=3.12.0
certifi~=2025.8.3
pymongo[srv]
pytest~=8.4.2
PyYAML~=6.0.2
setuptools~=80.9.0
scipy~=1.16.1